requirements.txt
supabase>=2.0.0
anthropic>=0.40.0
google-cloud-texttospeech>=2.16.0
google-cloud-storage>=2.14.0
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
